+++ config.h
@@ -119,6 +119,7 @@ struct config {
     char *font;
     bool use_bold;
     enum dpi_aware dpi_aware;
+    bool gamma_correct;
 
     uint16_t render_worker_count;
     uint16_t match_worker_count;