+++ icon.h
@@ -35,5 +35,5 @@ bool icon_lookup_application_icons(
     icon_theme_list_t themes, int icon_size,
     struct application_list *applications);
 
-bool icon_from_png(struct icon *icon, const char *name);
+bool icon_from_png(struct icon *icon, const char *name, bool gamma_correct);
 bool icon_from_svg(struct icon *icon, const char *name);