+++ shm.h
@@ -50,7 +50,9 @@ void shm_set_max_pool_size(off_t max_pool_size);
 
 struct buffer_chain;
 struct buffer_chain *shm_chain_new(
-    struct wl_shm *shm, bool scrollable, size_t pix_instances);
+    struct wl_shm *shm, bool scrollable, size_t pix_instances,
+    pixman_format_code_t pix_fmt_with_alpha,
+    pixman_format_code_t pix_fmt_without_alpha);
 void shm_chain_free(struct buffer_chain *chain);
 
 /*