+++ wayland.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#include <stdbool.h>
+#include <stdint.h>
+
 #include <fcft/fcft.h>
 
 #include "clipboard.h"
@@ -83,3 +86,5 @@ bool wayl_update_cache(const struct wayland *wayl);
 
 void wayl_clipboard_data(struct wayland *wayl, char *data, size_t size);
 void wayl_clipboard_done(struct wayland *wayl);
+
+bool wayl_do_linear_blending(const struct wayland *wayl);